third_party/blink/web_tests/virtual/partitioned-popins-disabled/external/wpt/partitioned-popins/partitioned-popins.cookies-blocked.tentative.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify Partitioned Popins cookie access when third-party cookie access blocked
  assert_equals: expected "ReadOnLoad:ThirdPartyNone-,ReadOnFetch:ThirdPartyNone-ThirdPartyNonePopin-,ReadOnDocument:ThirdPartyNone-ThirdPartyNonePopin-" but got "ReadOnLoad:FirstPartyLax-FirstPartyNone-,ReadOnFetch:FirstPartyStrict-FirstPartyLax-FirstPartyNone-FirstPartyStrictPopin-FirstPartyLaxPopin-FirstPartyNonePopin-ThirdPartyStrictPopin-ThirdPartyLaxPopin-ThirdPartyNonePopin-,ReadOnDocument:FirstPartyStrict-FirstPartyLax-FirstPartyNone-FirstPartyStrictPopin-FirstPartyLaxPopin-FirstPartyNonePopin-ThirdPartyStrictPopin-ThirdPartyLaxPopin-ThirdPartyNonePopin-"
Harness: the test ran to completion.
